* develop:
  fix bug: unable to create collection if it is root drive, eg. C:\
  Update subscribed state immediately
  Refactoring
  refresh personal cloud status from every 2 secs to 10 secs
  fix bug: progress incorrect if uploading more than Int32.Max
  Fix bug and adjust login/signup to v3
  fix bug and adjust layout
  fix bug: upload task count is not zeroed event if all done

+++ Source/Waveface.Stream.WindowsClient/Src/Class/ImportStatus.cs
@@ -57,8 +57,8 @@ namespace Waveface.Stream.WindowsClient
 	public class ImportStatusSummary
 	{
 		private string desc = "";
-		private int max;
-		private int cur;
+		private long max;
+		private long cur;
 		private bool hasProgress;
 
 		public ImportStatusSummary(ImportStatus status)
@@ -117,7 +117,7 @@ namespace Waveface.Stream.WindowsClient
 			get { return desc; }
 		}
 
-		public bool GetProgress(out int maximum, out int current)
+		public bool GetProgress(out long maximum, out long current)
 		{
 			maximum = max;
 			current = cur;
